requirements.txt
# NexusSentinel API Requirements
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
alembic==1.13.1
python-multipart==0.0.9
email-validator==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.27.0
tenacity==8.2.3
pandas==2.2.0
numpy==1.26.3
pytest==7.4.4
pytest-asyncio==0.23.5
gunicorn==21.2.0
psutil==5.9.8

